+++ app/service/send_notification.py
@@ -7,6 +7,7 @@ from app.dao.services_dao import dao_fetch_service_by_id
 from app.dao.templates_dao import dao_get_template_by_id_and_service_id
 from app.dao.users_dao import get_user_by_id
 from app.enums import KeyType, NotificationType, TemplateProcessType
+from app.errors import BadRequestError
 from app.notifications.process_notifications import (
     persist_notification,
     send_notification_to_queue,
@@ -16,7 +17,6 @@ from app.notifications.validators import (
     validate_and_format_recipient,
     validate_template,
 )
-from app.v2.errors import BadRequestError
 
 
 def validate_created_by(service, created_by_id):
